refactor: improve folder schema on sidebar-form
