+++ frontend/src/pages/private/workflows/workflows-editor/components/workflows-editor.component.tsx
@@ -13,7 +13,7 @@ import { toast } from "react-toastify"
 
 import * as yup from "yup"
 
-import { createInputsSchemaValidation } from './piece-form.component/validation';
+import { createInputsSchemaValidation } from './sidebar-form.component/piece-form.component/validation';
 import { yupResolver } from 'utils';
 import { storageFormSchema } from './sidebar-form.component/storage-form.component';
 import { ContainerResourceFormSchema } from './sidebar-form.component/container-resource-form.component';
